backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.35,<1.0
python-multipart>=0.0.20,<1.0
# Install the CUDA build with: pip install torch --index-url https://download.pytorch.org/whl/cu128
torch>=2.4,<3.0
torchvision>=0.19,<1.0
timm>=1.0,<2.0
Pillow>=11.0,<13.0
numpy>=2.2,<3.0
scikit-learn>=1.7,<2.0
python-dotenv>=1.0,<2.0
